Rename, refactor, some grouping changes
