+++ probafeladat/Functions/Cards/Model/CardPresentationModel.swift
@@ -1,0 +1,15 @@
+//
+//  CardPresentationModel.swift
+//  probafeladat
+//
+//  Created by Szabolcs Varga on 2020. 03. 19..
+//  Copyright © 2020. WUP. All rights reserved.
+//
+
+struct CardPresentationModel {
+    
+    let title: String
+    let curreny: String?
+    let value: String
+    
+}
